libpackage.so: change UrlProtocolListener size types to off_t
Missed this use after changing the definition in hrev54926

+++ src/kits/package/FetchFileJob.cpp
@@ -140,8 +140,8 @@ FetchFileJob::DataReceived(BUrlRequest*, const char* data, off_t position,
 
 
 void
-FetchFileJob::DownloadProgress(BUrlRequest*, ssize_t bytesReceived,
-	ssize_t bytesTotal)
+FetchFileJob::DownloadProgress(BUrlRequest*, off_t bytesReceived,
+	off_t bytesTotal)
 {
 	if (bytesTotal != 0) {
 		fBytes = bytesReceived;
