+++ src/kits/package/FetchFileJob.h
@@ -54,7 +54,7 @@ public:
 	virtual void	DataReceived(BUrlRequest*, const char* data,
 						off_t position, ssize_t size);
 	virtual void	DownloadProgress(BUrlRequest*,
-						ssize_t bytesReceived, ssize_t bytesTotal);
+						off_t bytesReceived, off_t bytesTotal);
 	virtual void 	RequestCompleted(BUrlRequest* request,
 						bool success);
 #endif
